refactor(programs): relocate GuardViolationError to its own guard module
Extract the shared guard-failure exception out of program.py into a
dedicated guard.py (PY-IC-9: exceptions in their own module). Both the
Program aggregate and the retry sub-machine raise it, so it belongs to
neither. Add GuardViolationError.reject as the single NoReturn raiser.

Narrow the programs package interface (PL-MD-3): GuardViolationError has
no consumer outside the package, so internal catchers import it from the
guard submodule rather than the package re-export.

program.py: module_size 300 -> 290, classes_per_module 2 -> 1.
